requirements.txt
pytest>=7.4.0
netmiko>=4.2.0
napalm>=4.1.0
pyyaml>=6.0
paramiko>=3.3.0
